Add base design files in commonMain
