+++ core/designsystem/src/iosMain/kotlin/com/droidknights/app/core/designsystem/theme/Theme.ios.kt
@@ -1,0 +1,8 @@
+package com.droidknights.app.core.designsystem.theme
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun WindowControl(darkTheme: Boolean) {
+    // NOT IMPLEMENTED
+}
